Measurement Window
New way to deal with the user selecting the charge carrier and mass
modifier. The change will allow the user to also save his selected
parameters which are then loaded (if the file exists) when the program
is started.

TODO: Rewrite the code to make it a bit cleaner if possible (and at the
very least remove the code duplication between close and save buttons).