Single file calibration fixes
Calibration on a single file would crash upon attempting to write the
calibrated file due to the xy library missing some imports. Furthermore,
the error file was being written to the MassyTools directory, I have
changed the code so that the error file is written to the directory
where the raw data file was picked from. This should be replaced at some
point by writing the error file to the same folder as where the user
puts the calibrated file.