Mass Error bug fix
The determination of the accurate mass after calibration gave a slightly
different value from what was listed in the error files. This was caused
due to the new x-axis that was used for the spline function being
different in both locations (the error file one used an automatically
scaling axis, while the extract data function used a fixed axis).