Tweaking & Bug fixing in mzML reader
The mzML reader was attempting to match file extension in case sensitive
way, this has been switched to case insensitive. Furthermore, the reader
now works with mzML files generated by the "t2dmzML" program, by
iterating through all 'scans' in a mzML file and returning the data of
the first 'scan' (as there should only be 1 scan to begin with).