MassyTools v1.0.0 b170724a
This build implements a new feature for automated data curation based on
parameters calculated by MassyTools (Issue #14) in a new menu item.
Furthermore, this build has moved the text that is presented in the
summary to a separate file (varnames) which makes it easier to
consistently change text that is shown to the users. Lastly, this build
has changed the indentation from tabs to spaces as per the PEP-8 (Issue
#3).