MassyTools v1.0.1 b180327a
This version implements an option to select different ways to determine
the background and noise as per issue #19. The current options are MIN,
MEDIAN or NOBAN, but these are currently hidden from the user in the
source code. This setting will be migrated to the GUI once it has been
extensively tested.

Furthermore, this version also adds a button in the Batch Process window
that allows users to choose if they want the program to overwrite any
existing analytes.ref file (which is required in case a user wants to
manually edit the analytes.ref file to omit a specific isotope).