MassyTools v1.0.2-alpha b180703a
This build addresses feature request #18, by parsing the calibrant line
and if the length equals one (meaning a line without an exact mass) than
MassyTools will calculate the m/z of the main isotope for the supplied
composition.

Furthermore, this build adds a building block for a reduced reducing
end, which proved troublesome when combined with permethylation, which
required a 'fake' oxygen molecule to be added to the composition to have
the correct number of permethylation sites be calculated by MassyTools.