MassyTools v0.1-alpha1 b181112a
This build implements functionality to calculate the isotopic pattern and mass of analytes using the defined 'blocks'. This build then uses this functionality in the calibration where the isotope that should be the highest isotope is used for calibration after providing a list of structures (this means that the mass is no longer needed in the calibration list).

This build still needs some tweaking to properly fragment the code that is currently in `calculate_analyte_isotopic_pattern` into several functions.