MassyTools v0.1-alpha1 b181206b
This build implements the minimum S/N for calibration as well as the minimum number of calibrants before actually calibrating a `MassSpectrum()`. However, it currently doesn't examine the region of the `MassSpectrum()` yet which refers to the `num_low_mass`, `num_medium_mass` and `num_high_mass` parameters of the settings.