MassyTools v0.1-alpha1 b190129a
This build implements the framework of the `Output` class, which
coincides with the renaming of the `Parameters` class to
`ProcessParameters` and the introduction of the `OutputParameters`
class. The current output that can be generated with this build is the
"Background Subtracted Intensities", which will serve as a prototype for
the other data types.