MassyTools v0.1-alpha1 b190131a
This build implements the option to change the experimental settings
using the `ExperimentalSettingsWindow`. Furthermore, this build fixes a
bug where the maximum_intensity of an isotope was not properly
quantified (bug was introduced in b190129d) and lastly, this build
introduces two new outputs (signal-to-noise and the non quantified
fraction).