MassyTools v0.1-alpha1 b190131b
This builds adds a numerical progress counter in the `SimpleProgressBar`
class and fixes a mistake in the residual fraction output, causing all
the values to be displayed inline without the measurement tags.

This build has uncovered that the progress bar doesn't behave as desired
as it doesn't take all steps into account, e.g. during calibration the
progress bar doesn't register the preperation time spent in determining
all the isotopes, but only starts running for the calibration itself).