MassyTools v0.1-alpha1 b190201d
This build implements the charge state functionality where the software
will attempt to calibrate on each charge state of the listed calibrants
during quantitation. Furthermore, the program will quantify each charge
state during quantitation and add the charge state between square
brackets behind the composition in the summary (for now).

Note: that if an analyte at a specific charge state falls outside the
measured m/z range, the software will print a WARNING to the log file
and ignore said analyte (in the given charge state).