MassyTools v0.1-alpha1 b190218a
This build contains several changes:

1. Some of the logic from 'MassyTools.py' towards the `MassSpectrum`
class, such as the quantitation.
2. This build implements more informative updates about what the
software is doing to the user by adding text to the GUI, above the
progress bar and each step has it's own progress bar attached to it.
3. This build also adds an option to get the 'Isotopic Pattern Quality'
as an output.
4. This build adds an option to select if the user wants PDF reports.
5. This build allows the user to save his default output selections.
6. This build ensures that if a new setting is added to the program, the
user can keep using their old settings files, while the program wil then
use the default value for the new settings.